Melchior FRANZ:
The current canopy opener/toggler has one drawback: it only works
when explicitly called. And that's only done in the keyboard definition.
So, one can engage the parking brake with Shift-B, and after a while
the canopy opens. But if one does now disengage the brakes with the
joystick (or in the property browser or wherever), then the canopy
won't be closed, which is quite unpleasant for the pilot, especially
in cold & wet weather.  ;-)   This cries for a listener. It watches
the parking brake property directly. Every change to it will operate
the canopy appropriately.