Alexis Bory: - Add blackout definitions so it matches what should feel a trained pilot. (same setting as the f-14b). - Remove the pilot's shoulder patch which was left blank. - Add the HSI model.
About the HSI:
At the moment, it just show magnetic heading and basic TACAN, that is *true* bearing and distance.
The course knob set the course property on nav[1] but it's not used for now.
The heading knob set instrumentation/heading-indicator-fg/offset-deg and is not used too.
Those two properties are probably not the good ones, this was just in order to check the animations.
Several objects are hidden, they are here (like the above un-used props) so Jorg can play with.